backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database clients
pymongo==4.6.0
redis==5.0.1
qdrant-client==1.7.0

# Job queue
rq==1.15.1

# ML and embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2
numpy==1.24.3
torch==2.1.1

# LLM and AI
openai>=1.13.3,<2.0.0
langchain>=0.1.10
langchain-community>=0.0.10
crewai==0.22.5

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Monitoring
prometheus-client==0.19.0